Guard against missing relations
An error started occurring around 2025-02-18T06:00:00Z (+/- a couple hours)
where the 'with relation as relation' statement on augmented_diff.py line 204
started raising an error. It might be that an empty relation was created? Or
that a previous OSMX update failed to insert a relation.

Anyways, adding this check unblocked the updater. More debugging needed.