Use rclone move instead of copy
This makes upload.sh upload files and immediately delete them, which
saves local disk space and also makes rclone faster since it doesn't
have to check a bunch of local files that have already been uploaded.

In order to support this we need to store a stamp file in stage-data/
so that Make can tell whether a changeset adiff's inputs have changed
since it was last built. I chose to store the md5 of each adiff file,
but an empty file would also work (Make only cares about the mtime).